refactor(pdf-service): remove circular dependencies
